chore: remove WebVitalsMonitor.tsx; archive legacy Web Vitals modules under _archive and adjust lint/format patterns
